vehicle-api/requirements.txt
django-ninja
kombu==5.2.*
